add score db , get user method

+++ server/spring-boot-jwt-JPA/src/main/java/com/javainuse/dao/UserDao.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.javainuse.model.DAOUser;
 
 @Repository
-public interface UserDao extends CrudRepository<DAOUser, Integer> {
+public interface UserDao extends CrudRepository<DAOUser, Long> {
 	
 	DAOUser findByUsername(String username);
